paint::pipeline_details

Struct PipelineDetails

pub(crate) struct PipelineDetails {

    pub pipeline: Option<CompositionPipeline>,
    pub parent_pipeline_id: Option<PipelineId>,
    pub children: Vec<PipelineId>,
    pub animations_running: bool,
    pub animation_callbacks_running: bool,
    pub throttled: bool,
    pub scroll_tree: ScrollTree,
    pub first_paint_metric: PaintMetricState,
    pub first_contentful_paint_metric: PaintMetricState,
    pub lcp_candidates: VecDeque<(Epoch, LCPCandidate)>,
    pub viewport_scale: Option<Scale<f32, CSSPixel, DevicePixel>>,
    pub exited: PipelineExitSource,
    pub display_list_epoch: Option<Epoch>,
    pub animations: PipelineAnimations,

}

Fields
pipeline: Option<CompositionPipeline>

The pipeline associated with this PipelineDetails object.

parent_pipeline_id: Option<PipelineId>

The id of the parent pipeline, if any.

children: Vec<PipelineId>

The ids of the child pipelines for this pipeline.

animations_running: bool

Whether animations are running

animation_callbacks_running: bool

Whether there are animation callbacks

throttled: bool

Whether to use less resources by stopping animations.

scroll_tree: ScrollTree

The Paint-side ScrollTree. This is used to allow finding and scrolling nodes in Paint before forwarding new offsets to WebRender.

first_paint_metric: PaintMetricState

The paint metric status of the first paint.

first_contentful_paint_metric: PaintMetricState

The paint metric status of the first contentful paint.

lcp_candidates: VecDeque<(Epoch, LCPCandidate)>

LCP candidates waiting to be presented, in order by WebRenderEpoch.

viewport_scale: Option<Scale<f32, CSSPixel, DevicePixel>>

The CSS pixel to device pixel scale of the viewport of this pipeline, including page zoom, but not including any pinch zoom amount. This is used to detect situations where the current display list is for an old scale.

exited: PipelineExitSource

Which parts of Servo have reported that this Pipeline has exited. Only when all have done so will it be discarded.

display_list_epoch: Option<Epoch>

The Epoch of the latest display list received for this Pipeline or None if no display list has been received.

animations: PipelineAnimations

Paint-driven animations associated with this PipelineDetails. Currently only text caret is handled this way.

Note: This does not manage animations and transitions from CSS or for user input interaction.

Implementations

impl PipelineDetails

pub(crate) fn animation_callbacks_running(&self) -> bool

pub(crate) fn animating(&self) -> bool

impl PipelineDetails

pub(crate) fn new() -> PipelineDetails

pub(crate) fn install_new_scroll_tree(&mut self, new_scroll_tree: ScrollTree)