paint

Module refresh_driver

Structs

AnimationRefreshDriverObserver 🔒

The AnimationRefreshDriverObserver is the default implementation of a RefreshDriver on systems without vsync hardware integration. It has a very basic way of triggering frames using a timer. It prevents new animation frames until the timer has fired.

BaseRefreshDriver 🔒

The BaseRefreshDriver is a “base class” for RefreshDriver trait implementations. It encapsulates shared behavior so that it does not have to be implemented by all trait implementations. It is responsible for providing RefreshDriver implementations with a callback that is used to wake up the event loop and trigger frame readiness.

TimerRefreshDriver 🔒

A thread that manages a TimerScheduler running in the background of the RefreshDriver. This is necessary because we need a reliable way of waking up the embedder’s main thread, which may just be sleeping until the EventLoopWaker asks it to wake up.

Enums

TimerThreadMessage 🔒

Traits

RefreshDriverObserver 🔒

A RefreshDriverObserver is an internal subscriber to frame start signals from the RefreshDriver. Examples of these kind of observers would be one that triggers new animation frames right after vsync signals or one that handles touch interactions once per frame.