paint

Module web_content_animation

Structs

CaretAnimation 🔒

Tracks the state of an ongoing caret blinking animation.

PipelineAnimations 🔒

This structure tracks the animations active for a given pipeline. Currently only caret blinking is tracked, but in the future this could perhaps track paint-side animations.

WebContentAnimator 🔒

A struct responsible for managing paint-side animations. Currently this only handles text caret blinking, but the idea is that in the future this would handle other types of paint-side animations as well.

Constants

CARET_BLINK_TIMEOUT 🔒

The amount of the time the caret blinks before ceasing, in order to preserve power. User activity (a new display list) will reset this.