paint

Module webview_renderer

Structs

ScrollEvent 🔒

ScrollResult 🔒

UnknownWebView

WebViewRenderer 🔒

A renderer for a libservo WebView. This is essentially the [ServoRenderer]’s interface to a libservo WebView, but the code here cannot depend on libservo in order to prevent circular dependencies, which is why we store a dyn WebViewTrait here instead of the WebView itself.

Enums

PinchZoomResult 🔒

ScrollZoomEvent 🔒